Refactor feature card  and add new Features component
