+++ src/frontend/components/features/index.ts
@@ -1,0 +1,2 @@
+export * from './features';
+export { default } from './features';
